src/main/java/com/watermelon/linkedlist/HasCycle.java
package com.watermelon.linkedlist;

/**
 * @author zhujiashuai
 * @date: 2024/2/17 19:45
 * @see : https://leetcode.cn/problems/linked-list-cycle/description
 */
public class HasCycle {

    public boolean hasCycle(ListNode head) {

        if (head == null || head.next == null) {
            return false;
        }

        ListNode slow = head;
        ListNode fast = head.next;

        while (slow != fast) {
            if (fast == null || fast.next == null) {
                return false;
            }
            slow = slow.next;
            fast = fast.next.next;
        }
        return true;
    }

    public static class ListNode {
        int val;
        ListNode next;

        ListNode() {
        }

        ListNode(int val) {
            this.val = val;
        }

        ListNode(int val, ListNode next) {
            this.val = val;
            this.next = next;
        }
    }

}
